making methods non static when possible, corrected bugs

+++ Database/src/main/java/fr/guiguilechat/eveonline/database/apiv2/Char.java
@@ -212,7 +212,7 @@ public class Char {
 	public static class OrderEntry {
 		public long orderID;
 		public long charID;
-		public int stationID;
+		public long stationID;
 		public int volEntered;
 		public int volRemaining;
 		public int minVolume;
